Remove html5slider Javascript helper
This helper was originally added as a "polyfill"
(backward-compatibility shim) for Firefox version 23 and older, so
that <input type="range"> that audio.js uses could render nicely in
those versions of Firefox.

Chromium did not need this shim, and Firefox 23 (released on Aug 6,
2013) added support for this. Therefore, let us remove this
dependency. By current metrics at caniuse.com
<http://caniuse.com/#feat=input-range>, 88.14% of web users support
this attribute. Opera Mini and old versions of Internet Explorer (6-8)
are the only holdouts.

Note also that the html5slider.js is uninterested in adding this to
the npm package manager (or any package manager) repository
<https://github.com/fryn/html5slider/issues/5>, meaning that keeping
it in the mediagoblin extlib section would mean manual
maintenance/updating by the GNU MediaGoblin team.

This commit also removes GMG's use of the shim.